requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
